new analysis: ListRemoteFiles
This is a "reporting" recipe which will output external files added as sources via ADD instructions.
This could be extended later to RUN command analysis, but I am still considering if I want to work
out some way to group "words" in the RUN command or just leave each word as a separate literal.