Remove usages of Lombok NonFinal which break equality
I found in my previous commit that the usage of Lombok's NonFinal attribute
would somehow break the generated equality which OpenRewrite relies on, causing
those recipes modifying the non-final variables to invoke 2 cycles.

Best practice is to stay single-cycle, so this updates those to remain single-cycle
with the hack of appying a Marker any time there's a minor LST change. Markers are still
useful to prevent expensive operations, though.