recipe: NameAllStages
This is a new recipe which explicitly names any unnamed stages by stage count,
excluding the final state. This is useful as numbered stages can lead to unintended
COPY if another intermediate stage is added which changes the stage numbers.

As a part of this, some details about how COPY options are processed are fixed.

Also fixes an issue in which multi-stage builds were duplicating stage children.